[web] PGF/TikZ export fixes
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@58445 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/export/pstricks/ExportSettings.java
@@ -20,9 +20,9 @@ public interface ExportSettings {
 
 	int getFormat();
 
-	int textYmaxValue();
+	double textYmaxValue();
 
-	int textYminValue();
+	double textYminValue();
 
 	void write(StringBuilder code);
